第 11 期：优化并发下单机制，提升用户体验和数据安全性 1. 实现分布式锁防止超卖：在 CartServiceImpl.submitOrder() 方法中集成 Redis 分布式锁机制，使用 RedisLockUtil 工具类对商品库存操作加锁，防止高并发场景下的库存超卖问题。 2. 优化库存扣减逻辑：改进 CartMapper.xml 中的 updateStore SQL 语句，使用 WHERE stock >= #{gshoppingnum} 悲观锁条件确保原子性操作，配合 @Transactional 事务注解实现失败回滚。 3. 解决金额精度问题：将 Order.amount、Goods.goprice、Goods.grprice 等金额字段从 Double 类型升级为 BigDecimal，避免浮点数计算导致的精度丢失（如20.999999998）；创建 MoneyUtil 工具类统一处理金额格式化和计算。 4. 优化支付交互体验：在 myOrder.html、pay.html 中改进支付流程，添加支付前确认对话框、处理中提示、成功/失败反馈等交互细节；修复支付按钮点击无响应的问题，确保每个操作都有明确的用户反馈。 5. 修复弹窗样式问题：在 custom.css 中新增 Element Plus 风格的对话框和提示框样式，规范图标大小（20px × 20px 和 48px × 48px），修复确认对话框中感叹号图标撑满整个页面的bug，提升视觉体验。

+++ src/main/java/com/ch/ebusiness/entity/Goods.java
@@ -1,12 +1,13 @@
 package com.ch.ebusiness.entity;
 
 import org.springframework.web.multipart.MultipartFile;
+import java.math.BigDecimal;
 
 public class Goods {
 	private int id;
 	private String gname;
-	private double goprice;
-	private double grprice;
+	private BigDecimal goprice;
+	private BigDecimal grprice;
 	private int stock;
 	private String gpicture;
 	private MultipartFile fileName;
@@ -31,19 +32,19 @@ public class Goods {
 		this.gname = gname;
 	}
 
-	public double getGoprice() {
+	public BigDecimal getGoprice() {
 		return goprice;
 	}
 
-	public void setGoprice(double goprice) {
+	public void setGoprice(BigDecimal goprice) {
 		this.goprice = goprice;
 	}
 
-	public double getGrprice() {
+	public BigDecimal getGrprice() {
 		return grprice;
 	}
 
-	public void setGrprice(double grprice) {
+	public void setGrprice(BigDecimal grprice) {
 		this.grprice = grprice;
 	}
